Add domain names in pf.conf to hosts
This way the gateways can come up with the correct firewall rules even
without any network connectivity. This was a problem during power
outages when the gateways would try and load the firewall before UTD's
nameservers came up. This would fail requiring a manual reboot later.

Hard coding the IPs should not be an issue as monit checks every minute
that the A records do not differ from what has be hard coded in hosts.