Firewall updates
update.freebsd.org added to both the nas and gateway firewalls to
ensure that the fileserver can update both for patches and for
major upgrades.

Neon is no longer a public facing box and so no longer requires
443 to anywhere, but requires 80 to anywhere due to the nature
of the raspbian mirror management.

There are no MS Windows domain controllers on the network
anymore, and thus no need to allow the DC's to talk to anything.
All ubuntu updates proxy through N anyway.