Fix ansible login for C
Because C must ask for a password to obtain a TGT for the user,
ansible was unable to login without also being provided an ssh
password.  This was undesireable as it broke the workflow for
pushing to groups with disimilar passwords.  This commit special
cases the local user to not go through the normal authentication
process and instead use only a keyed login.