Add Gutenburg to Icinga
As printers are likely to do, Gutenburg was found hiding between two functional
printers.  This common tactic is known to prevent the wrath of a user from
causing it harm, it also tends to prevent admins from tampering with the
insides of printers.  This shall happen no more as Gutenburg is now monitored
by icinga with the other printers and any outages shall now be noticed promptly.