Modify deployment strategies
This commit modifies the deployment strategy for clients to use the
newly available "free" strategy.  This should allow the clients to
advance through the playbook faster.

Additionally, the gateways are now configured to advance using a
serial strategy so that only one is modified at a time.  While this
has not been an issue in the past, we choose to be more cautious in
this particular case.