Use realm authentication for management of hosts
Instead of logging in to a shared root account, use normal
network provided auth to manage the vast majority of machines
on the network.  This commit also allows for group_vars to specify
additional groups that may be used to manage the machines.  Currently
this feature is used only on carbon to allow regular users to log in.