Commission new Koster and new Gutenburg
After 5+ years of service Koster and Gutenburg finally bit the dust
The two new printers fortunately use the same driver model, so its
only a change in the MAC address that must be made.  This commit
also updates the location of Franklin to match its physical
location with the other 2 printers.