Refactor Fileserver for Void
Initially it was thought that the FreeBSD system would be ideal for
use as a fileserver.  A year running it though has highlighted several
odd choices made by the FreeBSD maintainers.  Given that every
additional operating system that we run increases our "technical debt"
we have sought to refinance.  The move to Void has allowed us to
consolidate all of our technical debt to Void Linux instead of
continuing to operate a poorly understood FreeBSD system.