Create krb5_conf role
Previously the krb5.conf file was owned by the krb5-tools role.  When
this role was refactored into the netlogon role, there was no longer
any reason to have a seperate role, or so was thought at the time.  It
turns out that several other roles, notably the kdc itself, need that
file in place to bootstrap and should not have a cross dependency on
the netlogon role as it is not necessarily always desireable.