Auto-lock the default cv account on clients
The cv/cv login is used to troubleshoot a crashed ansible pull on
clients and to satisfy the requirements of creating a user account at
install time.  As it is not needed after the initial ansible-pull
completes it is automatically locked.  The cv account cannot be
automatically locked on servers as the deployment of server roles must
be done as push instead of a pull due to the established secrets
management procedures.  Servers are locked down after the initial
deployment has completed.