Deployment hotfix for pam_mount
This commit installs hxtools to allow the pam_mount.so stage to kill
user processes and to then unmount the home folders of network users.
While unconfirmed, it is proposed that the installation of this
package on oberon may have cleared a work queue and caused a cascading
systems failure on aluminium, beryllium, and boron.  This is
unconfirmed though is considered plausible.

Credit for this find goes to dinoocch.