Use static addressing
It has been shown that the dhcpcd IP address re-lease cycle causes the
fileserver to hang.  To solve this issue the addresses have been made
static to avoid the issue of lease updates.  This change has been
applied to all servers and will be rolled out to clients after it has
been shown to be stable.