Splitup Client Submodule
This commit splits up the client submodule.  Originally this module
was built with the intention of using a lower bar for quality when
judging commits so as to encourage more developers to work on it.  As
this has not been the case, and as managing the submodule is somewhat
inconvenient, the choice was made to remove the submodule.  All
functionality not directly driven by Collegium V requirements has been
derived from void-ansible-roles modules.  This is more inline with the
expected usage of submodules as seperate parts of a project that stay
locked at a particular version for an extended period of time.

This commit also moves the lightdm-webkit greeter out into a different
git repo.  This makes management of the greeter theme substantially
easier for the continued "Google Doodle" style changes that are
regularly made throughout the year.